Add exception for Apple macOS result code type
Apple version of pcsc-lite library is slightly different from official one,
and defines result code constants in pcsclite.h as unsigned 32bit integers.
Official pcsc-lite version uses LONG type and it is traslated to signed long on most platforms.
This implementation difference is solved using pre-processor instructions.

-----------------------------------------------------------------
Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

The contribution was created in whole or in part by me and I
have the right to submit it under the open source license
indicated in the file.

I understand and agree that this project and the contribution
are public and that a record of the contribution (including all
personal information I submit with it, including my sign-off) is
maintained indefinitely and may be redistributed consistent with
this project or the open source license(s) involved.

Signed-off-by: cyb3r4nt <104218001+cyb3r4nt@users.noreply.github.com>